+++ lib/base/game_component.dart
@@ -2,6 +2,7 @@ import 'dart:ui';
 
 import 'package:bonfire/base/bonfire_game.dart';
 import 'package:bonfire/collision/object_collision.dart';
+import 'package:bonfire/util/bonfire_game_ref.dart';
 import 'package:bonfire/util/interval_tick.dart';
 import 'package:bonfire/util/mixins/pointer_detector.dart';
 import 'package:bonfire/util/priority_layer.dart';
@@ -11,7 +12,7 @@ import 'package:flame/extensions.dart';
 
 /// Base of the all components in the Bonfire
 abstract class GameComponent extends Component
-    with HasGameRef<BonfireGame>, PointerDetectorHandler {
+    with BonfireHasGameRef<BonfireGame>, PointerDetectorHandler {
   /// Position used to draw on the screen
   Vector2Rect position = Vector2Rect.zero();