+++ lib/joystick/joystick_controller.dart
@@ -1,6 +1,5 @@
-import 'dart:ui';
-
 import 'package:bonfire/base/bonfire_game.dart';
+import 'package:bonfire/util/bonfire_game_ref.dart';
 import 'package:bonfire/util/mixins/pointer_detector.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:flame/components.dart';
@@ -58,7 +57,7 @@ mixin JoystickListener {
 }
 
 abstract class JoystickController extends Component
-    with HasGameRef<BonfireGame>, PointerDetectorHandler {
+    with BonfireHasGameRef<BonfireGame>, PointerDetectorHandler {
   List<JoystickListener> _observers = [];
   bool keyboardEnable = false;
 
@@ -91,12 +90,6 @@ abstract class JoystickController extends Component
   bool containObserver(JoystickListener listener) {
     return _observers.contains(listener);
   }
-
-  @override
-  void render(Canvas c) {}
-
-  @override
-  void update(double t) {}
 
   @override
   int get priority {