+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.2
+- Improvements in extensions organization.
+- Update Flame to `1.0.0-releasecandidate.13` version.
+- Replace `HasGameRef` for the own `BonfireHasGameRef`.
+
 # 1.3.1
 - little improvement in `drawDefaultLifeBar`.
 - create mixin `AutomaticRandomMovement`